+++ examples/expo-52/package.json
@@ -24,7 +24,7 @@
     "@rnx-kit/metro-config": "^2.0.1",
     "@rnx-kit/metro-resolver-symlinks": "^0.2.1",
     "@types/react": "~18.3.12",
-    "babel-preset-expo": "^12.0.11",
+    "babel-preset-expo": "12.0.11",
     "dotenv": "^16.4.5",
     "hot-updater": "workspace:*",
     "typescript": "^5.3.3"
